Fixes and updated caching service

+++ GamesDataService_API/Repository/GamesRepository.cs
@@ -17,13 +17,15 @@ namespace GamesDataService_API.Repository
     public class GamesRepository : IGamesRepository
     {
         private readonly IUploadService uploadService;
+        private readonly ICacheService cacheService;
         private readonly ApplicationDbContext db;
         private readonly IMapper mapper;
         private readonly IMemoryCache memoryCache;
 
-        public GamesRepository(IUploadService uploadService, ApplicationDbContext db, IMapper mapper, IMemoryCache memoryCache)
+        public GamesRepository(IUploadService uploadService, ICacheService cacheService ,ApplicationDbContext db, IMapper mapper, IMemoryCache memoryCache)
         {
             this.uploadService = uploadService;
+            this.cacheService = cacheService;
             this.db = db;
             this.mapper = mapper;
             this.memoryCache = memoryCache;
@@ -44,22 +46,26 @@ namespace GamesDataService_API.Repository
             mapped.IconId = results.PublicId;
 
             await db.Games.AddAsync(mapped);
-            if(await db.SaveChangesAsync() > 0) return new ResponseDto(true, StatusCodes.Status200OK, new[] { "Added game successfuly" });
+            if(await db.SaveChangesAsync() > 0){
+                await cacheService.TryRemoveFromCache<Games>(CacheType.Games);
+                return new ResponseDto(true, StatusCodes.Status200OK, new[] { "Added game successfuly" });
+            }
 
-            await uploadService.DeletePhoto(mapped.IconId);
-            await uploadService.DeletePhoto(mapped.CoverId);
+            if(!string.IsNullOrEmpty(mapped.IconId)) await uploadService.DeletePhoto(mapped.IconId);
+            if(!string.IsNullOrEmpty(mapped.CoverId)) await uploadService.DeletePhoto(mapped.CoverId);
             return new ResponseDto(false, StatusCodes.Status400BadRequest, new[] { "Could not add game" });
         }
 
         public async Task<ResponseDto> DeleteGame(Guid gameId)
         {
-            var game = await db.Games.FirstOrDefaultAsync(x => x.Id == gameId);
-            if(game is null) return new ResponseDto(false, StatusCodes.Status404NotFound, new[] { "Game does not exists" });
+            Games game;
+            if(!await GameExists(gameId, out game)) return new ResponseDto(false, StatusCodes.Status404NotFound, new[] { "Game does not exists" });
 
             db.Games.Remove(game);
             if(await db.SaveChangesAsync() > 0) {
-                await uploadService.DeletePhoto(game.IconId);
-                await uploadService.DeletePhoto(game.CoverId);
+                if(!string.IsNullOrEmpty(game.IconId)) await uploadService.DeletePhoto(game.IconId);
+                if(!string.IsNullOrEmpty(game.CoverId)) await uploadService.DeletePhoto(game.CoverId);
+                await cacheService.TryRemoveFromCache<Games>(CacheType.Games);
                 return new ResponseDto(true, StatusCodes.Status200OK, new[] { "Deleted game successfuly" });
             }
             return new ResponseDto(false, StatusCodes.Status400BadRequest, new[] { "Could not delete game" });
@@ -70,8 +76,8 @@ namespace GamesDataService_API.Repository
             var oldCoverId = string.Empty;
             var oldIconId = string.Empty;
 
-            var game = await db.Games.FirstOrDefaultAsync(x => x.Id == data.Id);
-            if(game is null) return new ResponseDto(false, StatusCodes.Status404NotFound, new[] { "Game does not exists" });
+            Games game;
+            if(!await GameExists(data.Id, out game)) return new ResponseDto(false, StatusCodes.Status404NotFound, new[] { "Game does not exists" });
 
             mapper.Map(data, game);
             if(data.CoverPhoto is not null){
@@ -92,7 +98,6 @@ namespace GamesDataService_API.Repository
                 game.IconUrl = results.SecureUrl.AbsoluteUri;
             }
             
-            db.Games.Update(game);
             if(await db.SaveChangesAsync() > 0)  {
                 if(!string.IsNullOrEmpty(oldIconId)) await uploadService.DeletePhoto(oldIconId); 
                 if(!string.IsNullOrEmpty(oldCoverId)) await uploadService.DeletePhoto(oldCoverId); 
@@ -106,16 +111,31 @@ namespace GamesDataService_API.Repository
 
         public async Task<ResponseDto> GetGames()
         {
-            IEnumerable<Games> games;
-            if(!memoryCache.TryGetValue("Games", out games)){
-                games = await db.Games.ToListAsync();
-
-                var cacheOptions = new MemoryCacheEntryOptions()
-                    .SetSlidingExpiration(TimeSpan.FromMinutes(5));
-
-                memoryCache.Set("Games", games, cacheOptions);
-            }
+            IEnumerable<Games> games = await cacheService.TryGetFromCache<Games>(CacheType.Games);
             return new ResponseDto(true, StatusCodes.Status200OK, mapper.Map<IEnumerable<GamesDto>>(games));
+        }
+
+        private Task<bool> GameExists(Guid gameId, out Games game){
+
+            IEnumerable<Games> gameExists;
+            if(memoryCache.TryGetValue(CacheType.Games, out gameExists)){
+                var findGame = gameExists.FirstOrDefault(x => x.Id == gameId);
+                if(findGame is not null) 
+                {
+                    game = findGame;
+                    return Task.FromResult(true);
+                }
+            }
+
+            var dbgameExists = db.Games.FirstOrDefault(x => x.Id == gameId);
+            if(dbgameExists is not null) 
+            {
+                game = dbgameExists;
+                return Task.FromResult(true);
+            }
+
+            game = new Games();
+            return Task.FromResult(false);
         }
     }
 }
